feature: power fragility displayer
Fragility curves specified in the power fragility data is now displayed under "Power Fragility" -> "Plot" tab.
